♻️ api: refactored GET /api/config
